Extract relative clauses and topic-chain linkage
