+++ main.js
@@ -13660,6 +13660,260 @@ var require_path_goal_source = __commonJS({
   }
 });
 
+// src/parser/topic-chain/linkage.js
+var require_linkage = __commonJS({
+  "src/parser/topic-chain/linkage.js"(exports2, module2) {
+    "use strict";
+    module2.exports = function createTopicChainLinkage2(dependencies = {}) {
+      const {
+        clauseSequenceSegments: clauseSequenceSegments2,
+        flattenNodes: flattenNodes2,
+        flattenSurface: flattenSurface2,
+        getConstructionAffordances: getConstructionAffordances2,
+        isClauseSequenceSeparator: isClauseSequenceSeparator2,
+        isRelationalCoverbLinker: isRelationalCoverbLinker2,
+        isTopicFrameLinker: isTopicFrameLinker2,
+        nodeCanFillSlot: nodeCanFillSlot2,
+        tokenSemanticDomains: tokenSemanticDomains2,
+        withoutIgnorableSpaceText: withoutIgnorableSpaceText2
+      } = dependencies;
+      function topicChainAntecedentDescriptor(segment = []) {
+        const rows = flattenNodes2(segment || []);
+        const constructions = rows.filter((row) => row.kind === "construction");
+        const npRows = constructions.filter((row) => {
+          const type = String(row.type || "");
+          return /NP$/u.test(type) || ["NominalHeadSpan", "ModifiedNP", "AssociativeNP", "PluralMarkedNP"].includes(type);
+        });
+        if (!npRows.length) return null;
+        const onlyMeaningful = withoutIgnorableSpaceText2(segment).filter((node) => !isClauseSequenceSeparator2(node));
+        const standaloneTopic = onlyMeaningful.length === 1 && onlyMeaningful[0].kind === "construction" && npRows.some((row) => row.surface === flattenSurface2(onlyMeaningful[0]));
+        const topicFrameLinker = onlyMeaningful.length >= 2 && isTopicFrameLinker2(onlyMeaningful[0]) ? onlyMeaningful[0] : null;
+        const relationalCoverbLinker = onlyMeaningful.length >= 2 && isRelationalCoverbLinker2(onlyMeaningful[0]) ? onlyMeaningful[0] : null;
+        const candidate = standaloneTopic ? npRows[0] : npRows[npRows.length - 1];
+        if (!candidate || !candidate.surface) return null;
+        const semanticDomains = Array.from(new Set(
+          rows.filter((row) => row.kind === "token").flatMap((row) => tokenSemanticDomains2(row))
+        ));
+        return {
+          surface: candidate.surface,
+          construction: candidate.type || "",
+          source: topicFrameLinker ? "overt_topic_frame" : standaloneTopic ? "overt_initial_topic" : "prior_clause_nominal_domain",
+          semantic_domains: semanticDomains,
+          topic_frame_linker_surface: topicFrameLinker ? topicFrameLinker.surface || "" : "",
+          topic_frame_status: topicFrameLinker ? "overt_topic_frame_linked" : "not_applicable",
+          relational_coverb_linker_surface: relationalCoverbLinker ? relationalCoverbLinker.surface || "" : "",
+          relational_frame_status: relationalCoverbLinker ? "overt_relational_coverb_accounted" : "not_applicable"
+        };
+      }
+      function visitConstructionNodes(node, visit, depth = 0) {
+        if (!node) return;
+        if (node.kind === "construction") {
+          visit(node, depth);
+          for (const child of node.children || []) visitConstructionNodes(child, visit, depth + 1);
+        }
+      }
+      function productiveObjectHeadToken2(node) {
+        const rows = flattenNodes2([node]);
+        return rows.find((row) => row.kind === "token" && (row.slots || []).includes("action_verb") && getConstructionAffordances2(row).can_head_productive_vo === true) || null;
+      }
+      function constructionHasOvertObject(node) {
+        const rows = flattenNodes2([node]);
+        const head = productiveObjectHeadToken2(node);
+        if (!head) return false;
+        const headIndex = rows.findIndex((row) => row.kind === "token" && row.surface === head.surface && row.depth === head.depth);
+        return rows.slice(Math.max(0, headIndex + 1)).some((row) => {
+          if (row.kind === "text") return false;
+          const slots = row.slots || [];
+          if (slots.includes("subject")) return false;
+          return slots.some((slot) => ["object", "theme", "head_noun", "np"].includes(slot));
+        });
+      }
+      function constructionContainsConstruction(root, target) {
+        if (!root || root.kind !== "construction") return false;
+        for (const child of root.children || []) {
+          if (child === target) return true;
+          if (child && child.kind === "construction" && constructionContainsConstruction(child, target)) return true;
+        }
+        return false;
+      }
+      function segmentHasOvertObjectAfterHead(segment = [], headSurface = "") {
+        const rows = flattenNodes2(segment || []);
+        const headIndex = rows.findIndex((row) => row.kind === "token" && row.surface === headSurface && (row.slots || []).includes("action_verb"));
+        if (headIndex < 0) return false;
+        return rows.slice(headIndex + 1).some((row) => {
+          if (row.kind === "text") return false;
+          const slots = row.slots || [];
+          if (slots.includes("subject") || slots.includes("particle")) return false;
+          return slots.some((slot) => ["object", "theme", "head_noun", "np"].includes(slot));
+        });
+      }
+      function topicChainGapCandidates(segment = []) {
+        const candidates = [];
+        for (const root of segment || []) {
+          visitConstructionNodes(root, (node, depth) => {
+            if (!nodeCanFillSlot2(node, "vp") && !nodeCanFillSlot2(node, "predicate")) return;
+            const head = productiveObjectHeadToken2(node);
+            if (!head || constructionHasOvertObject(node) || segmentHasOvertObjectAfterHead(segment, head.surface)) return;
+            candidates.push({ node, depth, head_surface: head.surface, surface: flattenSurface2(node) });
+          });
+        }
+        return candidates.filter((candidate) => !candidates.some(
+          (other) => other !== candidate && constructionContainsConstruction(candidate.node, other.node)
+        ));
+      }
+      function topicChainAntecedentCompatibility2(antecedent, head) {
+        const antecedentDomains = new Set(antecedent && antecedent.semantic_domains || []);
+        const syntax = String(head && head.syntax || "");
+        const result = (status, reason) => ({ status, reason, antecedent_domains: [...antecedentDomains] });
+        if (syntax.includes("chain_select_perception")) {
+          return result("compatible", "broad_perception_predicate_accepts_overt_nominal_topic");
+        }
+        if (syntax.includes("chain_select_discourse_content")) {
+          return result("compatible", "discourse_content_predicate_accepts_overt_topic_domain");
+        }
+        if (syntax.includes("chain_select_edible")) {
+          return antecedentDomains.has("food_item") || antecedentDomains.has("edible_item") ? result("compatible", "edible_topic_for_consumption_predicate") : antecedentDomains.size ? result("incompatible", "known_antecedent_domain_is_not_edible") : result("unverified", "eating_predicate_requires_known_edible_topic");
+        }
+        if (syntax.includes("chain_select_drinkable")) {
+          return antecedentDomains.has("drinkable_item") || antecedentDomains.has("liquid") ? result("compatible", "drinkable_topic_for_consumption_predicate") : antecedentDomains.size ? result("incompatible", "known_antecedent_domain_is_not_drinkable") : result("unverified", "drinking_predicate_requires_known_drinkable_topic");
+        }
+        if (syntax.includes("chain_select_bookable_resource")) {
+          return antecedentDomains.has("bookable_resource") ? result("compatible", "bookable_resource_topic_for_reservation_predicate") : result("unverified", "reservation_predicate_requires_known_bookable_resource_topic");
+        }
+        return result("unverified", "object_selecting_predicate_lacks_topic_chain_selection_profile");
+      }
+      function applyTopicChainNullObjectLinkage2(children = []) {
+        const segments = clauseSequenceSegments2(children);
+        if (segments.length < 2) return null;
+        const antecedent = topicChainAntecedentDescriptor(segments[0]);
+        if (!antecedent) return null;
+        const candidates = [];
+        for (const segment of segments.slice(1)) {
+          for (const candidate of topicChainGapCandidates(segment)) {
+            candidate.compatibility = topicChainAntecedentCompatibility2(antecedent, productiveObjectHeadToken2(candidate.node));
+            candidates.push(candidate);
+          }
+        }
+        if (!candidates.length) return null;
+        const blocked = candidates.filter((candidate) => candidate.compatibility.status !== "compatible");
+        const topicChainId = `topic-chain:${antecedent.surface}`;
+        const topicFrameTrace = antecedent.topic_frame_linker_surface ? {
+          topic_frame_status: antecedent.topic_frame_status,
+          topic_frame_linker_surface: antecedent.topic_frame_linker_surface,
+          topic_frame_domain_surface: antecedent.surface
+        } : {};
+        const relationalFrameTrace = antecedent.relational_coverb_linker_surface ? {
+          relational_frame_status: antecedent.relational_frame_status,
+          relational_coverb_linker_surface: antecedent.relational_coverb_linker_surface,
+          relational_coverb_domain_surface: antecedent.surface
+        } : {};
+        if (blocked.length) {
+          for (const candidate of candidates) {
+            const trace = candidate.node.trace || {};
+            candidate.node.trace = {
+              ...trace,
+              context_requirement_status: "context_required",
+              missing_argument_slots: ["object_or_activity_domain"],
+              missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "unresolved" }],
+              antecedent_status: candidate.compatibility.status === "incompatible" ? "incompatible" : "unverified",
+              antecedent_span: antecedent.surface,
+              discourse_license_not_observed: true,
+              topic_chain_id: topicChainId,
+              topic_chain_role: "unresolved_null_object_candidate",
+              topic_antecedent_surface: antecedent.surface,
+              topic_antecedent_source: antecedent.source,
+              ...topicFrameTrace,
+              ...relationalFrameTrace,
+              topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
+              null_object: "unresolved",
+              null_object_link: "not_licensed_antecedent_predicate_compatibility",
+              antecedent_predicate_compatibility_status: candidate.compatibility.status,
+              antecedent_predicate_compatibility_reason: candidate.compatibility.reason,
+              semantic_review_flags: Array.from(/* @__PURE__ */ new Set([...trace.semantic_review_flags || [], "topic_chain_antecedent_predicate_compatibility_review", "context_required_unresolved"])),
+              reason: "An overt topic is available, but the parser cannot license it as this predicate's omitted object unless their broad semantic domains are compatible. No hidden object is inserted.",
+              not_claims: Array.from(/* @__PURE__ */ new Set([...trace.not_claims || [], "not_fabricated_object_token", "not_unrestricted_argument_deletion", "not_semantically_incompatible_topic_link"]))
+            };
+          }
+          return {
+            topic_chain_id: topicChainId,
+            topic_chain_status: "antecedent_predicate_compatibility_review",
+            topic_antecedent_surface: antecedent.surface,
+            topic_antecedent_construction: antecedent.construction,
+            topic_antecedent_source: antecedent.source,
+            ...topicFrameTrace,
+            ...relationalFrameTrace,
+            topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
+            linked_null_object_count: 0,
+            linked_predicate_surfaces: [],
+            unresolved_predicate_surfaces: candidates.map((row) => row.surface),
+            unresolved_predicate_heads: candidates.map((row) => row.head_surface),
+            null_object_link: "not_licensed_antecedent_predicate_compatibility",
+            context_requirement_status: "context_required",
+            missing_argument_slots: ["object_or_activity_domain"],
+            missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "unresolved" }],
+            antecedent_status: blocked.some((row) => row.compatibility.status === "incompatible") ? "incompatible" : "unverified",
+            antecedent_span: antecedent.surface,
+            discourse_license_not_observed: true,
+            semantic_review_flags: ["topic_chain_antecedent_predicate_compatibility_review", "context_required_unresolved"],
+            not_claims: ["not_fabricated_object_token", "not_implicit_topic_inference", "not_unrestricted_argument_deletion", "not_semantically_incompatible_topic_link"]
+          };
+        }
+        for (const candidate of candidates) {
+          const trace = candidate.node.trace || {};
+          candidate.node.trace = {
+            ...trace,
+            context_requirement_status: "context_licensed",
+            missing_argument_slots: ["object_or_activity_domain"],
+            missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "licensed", licensed_by: topicChainId }],
+            antecedent_status: "linked",
+            antecedent_span: antecedent.surface,
+            discourse_license_not_observed: false,
+            topic_chain_id: topicChainId,
+            topic_chain_role: "null_object_target",
+            topic_antecedent_surface: antecedent.surface,
+            topic_antecedent_source: antecedent.source,
+            ...topicFrameTrace,
+            ...relationalFrameTrace,
+            topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
+            null_object: "licensed",
+            null_object_link: "licensed_to_overt_topic_chain",
+            antecedent_predicate_compatibility_status: "compatible",
+            antecedent_predicate_compatibility_reason: candidate.compatibility.reason,
+            reason: "An overt nominal topic/domain in the first linked segment licenses this later object-selecting predicate's unspoken object. The visible predicate remains unchanged and no hidden noun token is inserted.",
+            not_claims: Array.from(/* @__PURE__ */ new Set([...trace.not_claims || [], "not_fabricated_object_token", "not_context_free_object_drop", "not_unrestricted_argument_deletion"]))
+          };
+        }
+        return {
+          topic_chain_id: topicChainId,
+          topic_chain_status: "licensed_overt_topic_chain",
+          topic_antecedent_surface: antecedent.surface,
+          topic_antecedent_construction: antecedent.construction,
+          topic_antecedent_source: antecedent.source,
+          ...topicFrameTrace,
+          ...relationalFrameTrace,
+          topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
+          linked_null_object_count: candidates.length,
+          linked_predicate_surfaces: candidates.map((row) => row.surface),
+          linked_predicate_heads: candidates.map((row) => row.head_surface),
+          null_object_link: "licensed_to_overt_topic_chain",
+          context_requirement_status: "context_licensed",
+          missing_argument_slots: ["object_or_activity_domain"],
+          missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "licensed", licensed_by: topicChainId }],
+          antecedent_status: "linked",
+          antecedent_span: antecedent.surface,
+          discourse_license_not_observed: false,
+          not_claims: ["not_fabricated_object_token", "not_implicit_topic_inference", "not_unrestricted_argument_deletion"]
+        };
+      }
+      return {
+        applyTopicChainNullObjectLinkage: applyTopicChainNullObjectLinkage2,
+        productiveObjectHeadToken: productiveObjectHeadToken2,
+        topicChainAntecedentCompatibility: topicChainAntecedentCompatibility2
+      };
+    };
+  }
+});
+
 // src/parser/detectors/aspect/composition.js
 var require_composition2 = __commonJS({
   "src/parser/detectors/aspect/composition.js"(exports2, module2) {
@@ -16245,6 +16499,123 @@ var require_graph = __commonJS({
         shouldCollapseClauseSequenceForDisplay: shouldCollapseClauseSequenceForDisplay2,
         shouldCollapseGreedyWrapperForDisplay: shouldCollapseGreedyWrapperForDisplay2
       };
+    };
+  }
+});
+
+// src/parser/detectors/relative-clauses/clauses.js
+var require_clauses6 = __commonJS({
+  "src/parser/detectors/relative-clauses/clauses.js"(exports2, module2) {
+    "use strict";
+    module2.exports = function createRelativeClauseDetectors2(dependencies = {}) {
+      const {
+        applyConstructionPatterns: applyConstructionPatterns2,
+        clauseRelationParsedChunk: clauseRelationParsedChunk2,
+        clauseRelationSurfaceIndex: clauseRelationSurfaceIndex2,
+        cleanSlots: cleanSlots2,
+        construction: construction2,
+        cp021bIsBei2Marker: cp021bIsBei2Marker2,
+        flattenSurface: flattenSurface2,
+        nodeCanFillSlot: nodeCanFillSlot2,
+        parserInactiveTokenClone: parserInactiveTokenClone2,
+        traceInfo: traceInfo2,
+        withoutIgnorableSpaceText: withoutIgnorableSpaceText2
+      } = dependencies;
+      function relativeClauseGapPredicate(leftNodes = [], headSurface = "") {
+        const compact = withoutIgnorableSpaceText2(leftNodes || []);
+        if (compact.length !== 2 || !nodeCanFillSlot2(compact[0], "subject")) return null;
+        const verb = compact[1];
+        if (!nodeCanFillSlot2(verb, "action_verb")) return null;
+        const vp = construction2("TransitiveVP", "V+O", [verb], {
+          note: "Transitive predicate with an overt relative-clause head licensing the object gap; no hidden object token is inserted.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "TransitiveVP",
+            template_family: "generative_template",
+            template: ["action_verb!", "relative_object_gap!"],
+            assigned_slots: ["action_verb"],
+            relative_gap_status: "licensed_by_overt_head_noun",
+            relative_head_surface: headSurface,
+            missing_argument_slots: ["object"],
+            reason: "The relative-clause head noun supplies the overt nominal domain for the visible transitive predicate. The parser records the dependency without inserting an object token.",
+            not_claims: ["not_fabricated_object_token"]
+          })
+        });
+        return construction2("SubjectPredicateClause", "Clause", [compact[0], vp], {
+          note: "Relative modifier clause with overt subject and a head-licensed object gap.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "SubjectPredicateClause",
+            template_family: "generative_template",
+            template: ["subject!", "predicate!"],
+            assigned_slots: ["subject", "predicate"],
+            relative_gap_status: "licensed_by_overt_head_noun",
+            relative_head_surface: headSurface,
+            reason: "The overt subject and visible transitive predicate form the relative modifier clause; the following head noun licenses the object dependency.",
+            not_claims: ["not_fabricated_object_token"]
+          })
+        });
+      }
+      function relativeClauseNPForTerminal2(segment = []) {
+        const compact = withoutIgnorableSpaceText2(segment || []);
+        if (compact.filter(cp021bIsBei2Marker2).length >= 2) return null;
+        const linkerIndex = clauseRelationSurfaceIndex2(compact, ["嘅"]);
+        if (linkerIndex <= 0 || linkerIndex >= compact.length - 1) return null;
+        const head = compact[linkerIndex + 1];
+        if (!(nodeCanFillSlot2(head, "head_noun") || nodeCanFillSlot2(head, "time") || nodeCanFillSlot2(head, "time_head") || ["時候", "時間"].includes(flattenSurface2(head)))) return null;
+        const left = compact.slice(0, linkerIndex);
+        const headSurface = flattenSurface2(head);
+        const temporalHead = ["時候", "時間"].includes(headSurface);
+        const leftHasActionPredicate = left.some((node) => nodeCanFillSlot2(node, "action_verb"));
+        if (!temporalHead && !leftHasActionPredicate) return null;
+        const leftFinal = left[left.length - 1] || null;
+        const leftFinalLooksNominal = leftFinal && !nodeCanFillSlot2(leftFinal, "action_verb") && (nodeCanFillSlot2(leftFinal, "np") || nodeCanFillSlot2(leftFinal, "head_noun") || nodeCanFillSlot2(leftFinal, "location") || nodeCanFillSlot2(leftFinal, "object"));
+        if (!temporalHead && left.length >= 2 && leftFinalLooksNominal) return null;
+        let modifierParsed = applyConstructionPatterns2(left);
+        let modifierClause = modifierParsed.length === 1 && modifierParsed[0].kind === "construction" ? modifierParsed[0] : null;
+        if (!modifierClause) modifierClause = relativeClauseGapPredicate(left, headSurface);
+        if (!modifierClause) {
+          const chunk = clauseRelationParsedChunk2(left, { relative_head_surface: headSurface });
+          if (chunk.length === 1 && chunk[0].kind === "construction") modifierClause = chunk[0];
+        }
+        if (!modifierClause || !(nodeCanFillSlot2(modifierClause, "clause") || nodeCanFillSlot2(modifierClause, "predicate") || nodeCanFillSlot2(modifierClause, "vp"))) return null;
+        const linker = parserInactiveTokenClone2(compact[linkerIndex], {
+          label: "particle",
+          pos: "particle",
+          syntax: "nominal_linker relative_clause_linker",
+          slots: cleanSlots2([...compact[linkerIndex].slots || [], "nominal_linker", "relative_clause_linker"]),
+          reason: "嘅 links the visible modifier clause to its overt head noun inside a relative-clause NP."
+        });
+        const relativeNP = construction2("RelativeClauseNP", "RelNP", [modifierClause, linker, head], {
+          note: "Relative-clause noun phrase with a visible modifier clause, overt 嘅 linker, and overt head noun.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "RelativeClauseNP",
+            template_family: "generative_template",
+            template: ["relative_clause!", "nominal_linker!", "head_noun!"],
+            assigned_slots: ["relative_clause", "nominal_linker", "head_noun"],
+            relation_subtype: "relative_nominal_modifier",
+            relative_clause_construction: modifierClause.type,
+            head_noun_surface: headSurface,
+            temporal_head: ["時候", "時間"].includes(headSurface),
+            reason: "The modifier clause attaches inside the NP before any matrix predicate or temporal-subordinate relation is built.",
+            not_claims: ["not_flat_clause_sequence", "not_hidden_head_noun", "not_hidden_relative_gap_token"]
+          })
+        });
+        const tail = compact.slice(linkerIndex + 2);
+        if (!tail.length) return relativeNP;
+        const predicate = clauseRelationParsedChunk2(tail);
+        if (!predicate.length || !predicate.some((node) => node && (node.kind === "construction" || nodeCanFillSlot2(node, "predicate") || nodeCanFillSlot2(node, "vp")))) return null;
+        return construction2("SubjectPredicateClause", "Clause", [relativeNP, ...predicate], {
+          note: "Matrix clause whose overt subject/topic is a relative-clause NP.",
+          trace: traceInfo2("generative_template", {
+            construction_type: "SubjectPredicateClause",
+            template_family: "generative_template",
+            template: ["subject!", "predicate!"],
+            assigned_slots: ["subject", "predicate"],
+            relative_np_subject: true,
+            reason: "The relative-clause NP is completed before attachment to the visible matrix predicate."
+          })
+        });
+      }
+      return { relativeClauseNPForTerminal: relativeClauseNPForTerminal2 };
     };
   }
 });
@@ -19016,6 +19387,23 @@ function tokenSemanticDomains(node) {
   const bundle = tokenNode && (tokenNode.feature_bundle || tokenNode.trace && tokenNode.trace.feature_bundle);
   return bundle && bundle.parser_features && Array.isArray(bundle.parser_features.semantic_domain) ? bundle.parser_features.semantic_domain : [];
 }
+var createTopicChainLinkage = require_linkage();
+var {
+  applyTopicChainNullObjectLinkage,
+  productiveObjectHeadToken,
+  topicChainAntecedentCompatibility
+} = createTopicChainLinkage({
+  clauseSequenceSegments: (...args) => clauseSequenceSegments(...args),
+  flattenNodes,
+  flattenSurface,
+  getConstructionAffordances,
+  isClauseSequenceSeparator,
+  isRelationalCoverbLinker,
+  isTopicFrameLinker,
+  nodeCanFillSlot,
+  tokenSemanticDomains,
+  withoutIgnorableSpaceText
+});
 var OVERT_OBJECT_SELECTION_REVIEW_TYPES = /* @__PURE__ */ new Set(["PerfectiveVP", "PostverbalZoPerfectiveVP", "TransitiveVP", "CompletionVP", "ProductiveVO"]);
 function assignedConstructionChild(node, slotNames = []) {
   if (!node || node.kind !== "construction") return null;
@@ -19737,328 +20125,6 @@ function wrapCore(core) {
   if (modalPredicateWrapSpan) return modalPredicateWrapSpan;
   return wrapPredicate(core);
 }
-function relativeClauseGapPredicate(leftNodes = [], headSurface = "") {
-  const compact = withoutIgnorableSpaceText(leftNodes || []);
-  if (compact.length !== 2 || !nodeCanFillSlot(compact[0], "subject")) return null;
-  const verb = compact[1];
-  if (!nodeCanFillSlot(verb, "action_verb")) return null;
-  const vp = construction("TransitiveVP", "V+O", [verb], {
-    note: "Transitive predicate with an overt relative-clause head licensing the object gap; no hidden object token is inserted.",
-    trace: traceInfo("generative_template", {
-      construction_type: "TransitiveVP",
-      template_family: "generative_template",
-      template: ["action_verb!", "relative_object_gap!"],
-      assigned_slots: ["action_verb"],
-      relative_gap_status: "licensed_by_overt_head_noun",
-      relative_head_surface: headSurface,
-      missing_argument_slots: ["object"],
-      reason: "The relative-clause head noun supplies the overt nominal domain for the visible transitive predicate. The parser records the dependency without inserting an object token.",
-      not_claims: ["not_fabricated_object_token"]
-    })
-  });
-  return construction("SubjectPredicateClause", "Clause", [compact[0], vp], {
-    note: "Relative modifier clause with overt subject and a head-licensed object gap.",
-    trace: traceInfo("generative_template", {
-      construction_type: "SubjectPredicateClause",
-      template_family: "generative_template",
-      template: ["subject!", "predicate!"],
-      assigned_slots: ["subject", "predicate"],
-      relative_gap_status: "licensed_by_overt_head_noun",
-      relative_head_surface: headSurface,
-      reason: "The overt subject and visible transitive predicate form the relative modifier clause; the following head noun licenses the object dependency.",
-      not_claims: ["not_fabricated_object_token"]
-    })
-  });
-}
-function relativeClauseNPForTerminal(segment = []) {
-  const compact = withoutIgnorableSpaceText(segment || []);
-  if (compact.filter(cp021bIsBei2Marker).length >= 2) return null;
-  const linkerIndex = clauseRelationSurfaceIndex(compact, ["嘅"]);
-  if (linkerIndex <= 0 || linkerIndex >= compact.length - 1) return null;
-  const head = compact[linkerIndex + 1];
-  if (!(nodeCanFillSlot(head, "head_noun") || nodeCanFillSlot(head, "time") || nodeCanFillSlot(head, "time_head") || ["時候", "時間"].includes(flattenSurface(head)))) return null;
-  const left = compact.slice(0, linkerIndex);
-  const headSurface = flattenSurface(head);
-  const temporalHead = ["時候", "時間"].includes(headSurface);
-  const leftHasActionPredicate = left.some((node) => nodeCanFillSlot(node, "action_verb"));
-  if (!temporalHead && !leftHasActionPredicate) return null;
-  const leftFinal = left[left.length - 1] || null;
-  const leftFinalLooksNominal = leftFinal && !nodeCanFillSlot(leftFinal, "action_verb") && (nodeCanFillSlot(leftFinal, "np") || nodeCanFillSlot(leftFinal, "head_noun") || nodeCanFillSlot(leftFinal, "location") || nodeCanFillSlot(leftFinal, "object"));
-  if (!temporalHead && left.length >= 2 && leftFinalLooksNominal) return null;
-  let modifierParsed = applyConstructionPatterns(left);
-  let modifierClause = modifierParsed.length === 1 && modifierParsed[0].kind === "construction" ? modifierParsed[0] : null;
-  if (!modifierClause) modifierClause = relativeClauseGapPredicate(left, headSurface);
-  if (!modifierClause) {
-    const chunk = clauseRelationParsedChunk(left, { relative_head_surface: headSurface });
-    if (chunk.length === 1 && chunk[0].kind === "construction") modifierClause = chunk[0];
-  }
-  if (!modifierClause || !(nodeCanFillSlot(modifierClause, "clause") || nodeCanFillSlot(modifierClause, "predicate") || nodeCanFillSlot(modifierClause, "vp"))) return null;
-  const linker = parserInactiveTokenClone(compact[linkerIndex], {
-    label: "particle",
-    pos: "particle",
-    syntax: "nominal_linker relative_clause_linker",
-    slots: cleanSlots([...compact[linkerIndex].slots || [], "nominal_linker", "relative_clause_linker"]),
-    reason: "嘅 links the visible modifier clause to its overt head noun inside a relative-clause NP."
-  });
-  const relativeNP = construction("RelativeClauseNP", "RelNP", [modifierClause, linker, head], {
-    note: "Relative-clause noun phrase with a visible modifier clause, overt 嘅 linker, and overt head noun.",
-    trace: traceInfo("generative_template", {
-      construction_type: "RelativeClauseNP",
-      template_family: "generative_template",
-      template: ["relative_clause!", "nominal_linker!", "head_noun!"],
-      assigned_slots: ["relative_clause", "nominal_linker", "head_noun"],
-      relation_subtype: "relative_nominal_modifier",
-      relative_clause_construction: modifierClause.type,
-      head_noun_surface: headSurface,
-      temporal_head: ["時候", "時間"].includes(headSurface),
-      reason: "The modifier clause attaches inside the NP before any matrix predicate or temporal-subordinate relation is built.",
-      not_claims: ["not_flat_clause_sequence", "not_hidden_head_noun", "not_hidden_relative_gap_token"]
-    })
-  });
-  const tail = compact.slice(linkerIndex + 2);
-  if (!tail.length) return relativeNP;
-  const predicate = clauseRelationParsedChunk(tail);
-  if (!predicate.length || !predicate.some((node) => node && (node.kind === "construction" || nodeCanFillSlot(node, "predicate") || nodeCanFillSlot(node, "vp")))) return null;
-  return construction("SubjectPredicateClause", "Clause", [relativeNP, ...predicate], {
-    note: "Matrix clause whose overt subject/topic is a relative-clause NP.",
-    trace: traceInfo("generative_template", {
-      construction_type: "SubjectPredicateClause",
-      template_family: "generative_template",
-      template: ["subject!", "predicate!"],
-      assigned_slots: ["subject", "predicate"],
-      relative_np_subject: true,
-      reason: "The relative-clause NP is completed before attachment to the visible matrix predicate."
-    })
-  });
-}
-function topicChainAntecedentDescriptor(segment = []) {
-  const rows = flattenNodes(segment || []);
-  const constructions = rows.filter((row) => row.kind === "construction");
-  const npRows = constructions.filter((row) => {
-    const type = String(row.type || "");
-    return /NP$/u.test(type) || ["NominalHeadSpan", "ModifiedNP", "AssociativeNP", "PluralMarkedNP"].includes(type);
-  });
-  if (!npRows.length) return null;
-  const onlyMeaningful = withoutIgnorableSpaceText(segment).filter((node) => !isClauseSequenceSeparator(node));
-  const standaloneTopic = onlyMeaningful.length === 1 && onlyMeaningful[0].kind === "construction" && npRows.some((row) => row.surface === flattenSurface(onlyMeaningful[0]));
-  const topicFrameLinker = onlyMeaningful.length >= 2 && isTopicFrameLinker(onlyMeaningful[0]) ? onlyMeaningful[0] : null;
-  const relationalCoverbLinker = onlyMeaningful.length >= 2 && isRelationalCoverbLinker(onlyMeaningful[0]) ? onlyMeaningful[0] : null;
-  const candidate = standaloneTopic ? npRows[0] : npRows[npRows.length - 1];
-  if (!candidate || !candidate.surface) return null;
-  const semanticDomains = Array.from(new Set(
-    rows.filter((row) => row.kind === "token").flatMap((row) => tokenSemanticDomains(row))
-  ));
-  return {
-    surface: candidate.surface,
-    construction: candidate.type || "",
-    source: topicFrameLinker ? "overt_topic_frame" : standaloneTopic ? "overt_initial_topic" : "prior_clause_nominal_domain",
-    semantic_domains: semanticDomains,
-    topic_frame_linker_surface: topicFrameLinker ? topicFrameLinker.surface || "" : "",
-    topic_frame_status: topicFrameLinker ? "overt_topic_frame_linked" : "not_applicable",
-    relational_coverb_linker_surface: relationalCoverbLinker ? relationalCoverbLinker.surface || "" : "",
-    relational_frame_status: relationalCoverbLinker ? "overt_relational_coverb_accounted" : "not_applicable"
-  };
-}
-function visitConstructionNodes(node, visit, depth = 0) {
-  if (!node) return;
-  if (node.kind === "construction") {
-    visit(node, depth);
-    for (const child of node.children || []) visitConstructionNodes(child, visit, depth + 1);
-  }
-}
-function productiveObjectHeadToken(node) {
-  const rows = flattenNodes([node]);
-  return rows.find((row) => row.kind === "token" && (row.slots || []).includes("action_verb") && getConstructionAffordances(row).can_head_productive_vo === true) || null;
-}
-function constructionHasOvertObject(node) {
-  const rows = flattenNodes([node]);
-  const head = productiveObjectHeadToken(node);
-  if (!head) return false;
-  const headIndex = rows.findIndex((row) => row.kind === "token" && row.surface === head.surface && row.depth === head.depth);
-  return rows.slice(Math.max(0, headIndex + 1)).some((row) => {
-    if (row.kind === "text") return false;
-    const slots = row.slots || [];
-    if (slots.includes("subject")) return false;
-    return slots.some((slot) => ["object", "theme", "head_noun", "np"].includes(slot));
-  });
-}
-function constructionContainsConstruction(root, target) {
-  if (!root || root.kind !== "construction") return false;
-  for (const child of root.children || []) {
-    if (child === target) return true;
-    if (child && child.kind === "construction" && constructionContainsConstruction(child, target)) return true;
-  }
-  return false;
-}
-function segmentHasOvertObjectAfterHead(segment = [], headSurface = "") {
-  const rows = flattenNodes(segment || []);
-  const headIndex = rows.findIndex((row) => row.kind === "token" && row.surface === headSurface && (row.slots || []).includes("action_verb"));
-  if (headIndex < 0) return false;
-  return rows.slice(headIndex + 1).some((row) => {
-    if (row.kind === "text") return false;
-    const slots = row.slots || [];
-    if (slots.includes("subject") || slots.includes("particle")) return false;
-    return slots.some((slot) => ["object", "theme", "head_noun", "np"].includes(slot));
-  });
-}
-function topicChainGapCandidates(segment = []) {
-  const candidates = [];
-  for (const root of segment || []) {
-    visitConstructionNodes(root, (node, depth) => {
-      if (!nodeCanFillSlot(node, "vp") && !nodeCanFillSlot(node, "predicate")) return;
-      const head = productiveObjectHeadToken(node);
-      if (!head || constructionHasOvertObject(node) || segmentHasOvertObjectAfterHead(segment, head.surface)) return;
-      candidates.push({ node, depth, head_surface: head.surface, surface: flattenSurface(node) });
-    });
-  }
-  return candidates.filter((candidate) => !candidates.some(
-    (other) => other !== candidate && constructionContainsConstruction(candidate.node, other.node)
-  ));
-}
-function topicChainAntecedentCompatibility(antecedent, head) {
-  const antecedentDomains = new Set(antecedent && antecedent.semantic_domains || []);
-  const syntax = String(head && head.syntax || "");
-  const result = (status, reason) => ({ status, reason, antecedent_domains: [...antecedentDomains] });
-  if (syntax.includes("chain_select_perception")) {
-    return result("compatible", "broad_perception_predicate_accepts_overt_nominal_topic");
-  }
-  if (syntax.includes("chain_select_discourse_content")) {
-    return result("compatible", "discourse_content_predicate_accepts_overt_topic_domain");
-  }
-  if (syntax.includes("chain_select_edible")) {
-    return antecedentDomains.has("food_item") || antecedentDomains.has("edible_item") ? result("compatible", "edible_topic_for_consumption_predicate") : antecedentDomains.size ? result("incompatible", "known_antecedent_domain_is_not_edible") : result("unverified", "eating_predicate_requires_known_edible_topic");
-  }
-  if (syntax.includes("chain_select_drinkable")) {
-    return antecedentDomains.has("drinkable_item") || antecedentDomains.has("liquid") ? result("compatible", "drinkable_topic_for_consumption_predicate") : antecedentDomains.size ? result("incompatible", "known_antecedent_domain_is_not_drinkable") : result("unverified", "drinking_predicate_requires_known_drinkable_topic");
-  }
-  if (syntax.includes("chain_select_bookable_resource")) {
-    return antecedentDomains.has("bookable_resource") ? result("compatible", "bookable_resource_topic_for_reservation_predicate") : result("unverified", "reservation_predicate_requires_known_bookable_resource_topic");
-  }
-  return result("unverified", "object_selecting_predicate_lacks_topic_chain_selection_profile");
-}
-function applyTopicChainNullObjectLinkage(children = []) {
-  const segments = clauseSequenceSegments(children);
-  if (segments.length < 2) return null;
-  const antecedent = topicChainAntecedentDescriptor(segments[0]);
-  if (!antecedent) return null;
-  const candidates = [];
-  for (const segment of segments.slice(1)) {
-    for (const candidate of topicChainGapCandidates(segment)) {
-      candidate.compatibility = topicChainAntecedentCompatibility(antecedent, productiveObjectHeadToken(candidate.node));
-      candidates.push(candidate);
-    }
-  }
-  if (!candidates.length) return null;
-  const blocked = candidates.filter((candidate) => candidate.compatibility.status !== "compatible");
-  const topicChainId = `topic-chain:${antecedent.surface}`;
-  const topicFrameTrace = antecedent.topic_frame_linker_surface ? {
-    topic_frame_status: antecedent.topic_frame_status,
-    topic_frame_linker_surface: antecedent.topic_frame_linker_surface,
-    topic_frame_domain_surface: antecedent.surface
-  } : {};
-  const relationalFrameTrace = antecedent.relational_coverb_linker_surface ? {
-    relational_frame_status: antecedent.relational_frame_status,
-    relational_coverb_linker_surface: antecedent.relational_coverb_linker_surface,
-    relational_coverb_domain_surface: antecedent.surface
-  } : {};
-  if (blocked.length) {
-    for (const candidate of candidates) {
-      const trace = candidate.node.trace || {};
-      candidate.node.trace = {
-        ...trace,
-        context_requirement_status: "context_required",
-        missing_argument_slots: ["object_or_activity_domain"],
-        missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "unresolved" }],
-        antecedent_status: candidate.compatibility.status === "incompatible" ? "incompatible" : "unverified",
-        antecedent_span: antecedent.surface,
-        discourse_license_not_observed: true,
-        topic_chain_id: topicChainId,
-        topic_chain_role: "unresolved_null_object_candidate",
-        topic_antecedent_surface: antecedent.surface,
-        topic_antecedent_source: antecedent.source,
-        ...topicFrameTrace,
-        ...relationalFrameTrace,
-        topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
-        null_object: "unresolved",
-        null_object_link: "not_licensed_antecedent_predicate_compatibility",
-        antecedent_predicate_compatibility_status: candidate.compatibility.status,
-        antecedent_predicate_compatibility_reason: candidate.compatibility.reason,
-        semantic_review_flags: Array.from(/* @__PURE__ */ new Set([...trace.semantic_review_flags || [], "topic_chain_antecedent_predicate_compatibility_review", "context_required_unresolved"])),
-        reason: "An overt topic is available, but the parser cannot license it as this predicate's omitted object unless their broad semantic domains are compatible. No hidden object is inserted.",
-        not_claims: Array.from(/* @__PURE__ */ new Set([...trace.not_claims || [], "not_fabricated_object_token", "not_unrestricted_argument_deletion", "not_semantically_incompatible_topic_link"]))
-      };
-    }
-    return {
-      topic_chain_id: topicChainId,
-      topic_chain_status: "antecedent_predicate_compatibility_review",
-      topic_antecedent_surface: antecedent.surface,
-      topic_antecedent_construction: antecedent.construction,
-      topic_antecedent_source: antecedent.source,
-      ...topicFrameTrace,
-      ...relationalFrameTrace,
-      topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
-      linked_null_object_count: 0,
-      linked_predicate_surfaces: [],
-      unresolved_predicate_surfaces: candidates.map((row) => row.surface),
-      unresolved_predicate_heads: candidates.map((row) => row.head_surface),
-      null_object_link: "not_licensed_antecedent_predicate_compatibility",
-      context_requirement_status: "context_required",
-      missing_argument_slots: ["object_or_activity_domain"],
-      missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "unresolved" }],
-      antecedent_status: blocked.some((row) => row.compatibility.status === "incompatible") ? "incompatible" : "unverified",
-      antecedent_span: antecedent.surface,
-      discourse_license_not_observed: true,
-      semantic_review_flags: ["topic_chain_antecedent_predicate_compatibility_review", "context_required_unresolved"],
-      not_claims: ["not_fabricated_object_token", "not_implicit_topic_inference", "not_unrestricted_argument_deletion", "not_semantically_incompatible_topic_link"]
-    };
-  }
-  for (const candidate of candidates) {
-    const trace = candidate.node.trace || {};
-    candidate.node.trace = {
-      ...trace,
-      context_requirement_status: "context_licensed",
-      missing_argument_slots: ["object_or_activity_domain"],
-      missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "licensed", licensed_by: topicChainId }],
-      antecedent_status: "linked",
-      antecedent_span: antecedent.surface,
-      discourse_license_not_observed: false,
-      topic_chain_id: topicChainId,
-      topic_chain_role: "null_object_target",
-      topic_antecedent_surface: antecedent.surface,
-      topic_antecedent_source: antecedent.source,
-      ...topicFrameTrace,
-      ...relationalFrameTrace,
-      topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
-      null_object: "licensed",
-      null_object_link: "licensed_to_overt_topic_chain",
-      antecedent_predicate_compatibility_status: "compatible",
-      antecedent_predicate_compatibility_reason: candidate.compatibility.reason,
-      reason: "An overt nominal topic/domain in the first linked segment licenses this later object-selecting predicate's unspoken object. The visible predicate remains unchanged and no hidden noun token is inserted.",
-      not_claims: Array.from(/* @__PURE__ */ new Set([...trace.not_claims || [], "not_fabricated_object_token", "not_context_free_object_drop", "not_unrestricted_argument_deletion"]))
-    };
-  }
-  return {
-    topic_chain_id: topicChainId,
-    topic_chain_status: "licensed_overt_topic_chain",
-    topic_antecedent_surface: antecedent.surface,
-    topic_antecedent_construction: antecedent.construction,
-    topic_antecedent_source: antecedent.source,
-    ...topicFrameTrace,
-    ...relationalFrameTrace,
-    topic_antecedent_semantic_domains: antecedent.semantic_domains || [],
-    linked_null_object_count: candidates.length,
-    linked_predicate_surfaces: candidates.map((row) => row.surface),
-    linked_predicate_heads: candidates.map((row) => row.head_surface),
-    null_object_link: "licensed_to_overt_topic_chain",
-    context_requirement_status: "context_licensed",
-    missing_argument_slots: ["object_or_activity_domain"],
-    missing_slot_details: [{ slot: "object_or_activity_domain", license_status: "licensed", licensed_by: topicChainId }],
-    antecedent_status: "linked",
-    antecedent_span: antecedent.surface,
-    discourse_license_not_observed: false,
-    not_claims: ["not_fabricated_object_token", "not_implicit_topic_inference", "not_unrestricted_argument_deletion"]
-  };
-}
 var createClauseRelationGraph = require_graph();
 var {
   connectorAwareClauseLinkingForTerminal,
@@ -20095,6 +20161,20 @@ var {
   reportedSpeechFrameFallback,
   surfaceOf,
   token: token2,
+  traceInfo,
+  withoutIgnorableSpaceText
+});
+var createRelativeClauseDetectors = require_clauses6();
+var { relativeClauseNPForTerminal } = createRelativeClauseDetectors({
+  applyConstructionPatterns,
+  clauseRelationParsedChunk,
+  clauseRelationSurfaceIndex,
+  cleanSlots,
+  construction,
+  cp021bIsBei2Marker,
+  flattenSurface,
+  nodeCanFillSlot,
+  parserInactiveTokenClone,
   traceInfo,
   withoutIgnorableSpaceText
 });
